refactor: extract ChatHeader component, fix keyboard flicker on mobile
- Extract chat header (~250 lines) into ChatHeader.tsx — pure display
  component with back button, squad name, member avatars, date badges,
  expiry countdown, and +7 days extension
- Fix mobile keyboard flicker: replace chatHeight state with direct DOM
  mutation via ref so viewport resize events no longer trigger re-renders
